+++ grid2op/Opponent/geometricOpponent.py
@@ -11,7 +11,7 @@ import copy
 import numpy as np
 
 from grid2op.dtypes import dt_int
-from grid2op.Opponent import BaseOpponent
+from grid2op.Opponent.baseOpponent import BaseOpponent
 from grid2op.Exceptions import OpponentError